ai/litellm/requirements.txt
litellm>=1.62.1
python-dotenv>=1.0.0
